requirements/base.txt

Django==2.0.5

Markdown==2.6.11

Underscore==0.0.30000001

also==0.0.2

dj-database-url==0.5.0

django-bower==5.2.0

django-crispy-forms==1.7.2

django-environ==0.4.4

django-extensions==2.0.7

django-filter==1.1.0

django-pipeline==1.6.14

djangorestframework==3.8.2

pytz==2018.4

six==1.11.0

sqlparse==0.2.4

https://github.com/CMCuritiba/django-cmcldapauth/raw/master/dist/django-cmcldapauth-0.2.tar.gz

django-python3-ldap==0.11.1
django-ldapdb==1.0.0

django-braces==1.13.0

simplejson==3.15.0

factory-boy==2.11.1
